moving the usize u32 conversion to a lower level

+++ src/commands/test_balances.rs
@@ -9,13 +9,13 @@ use zcash_primitives::transaction::Transaction;
 pub(crate) struct TestBalances(Vec<u64>);
 
 impl TestBalances {
-    pub(crate) fn get_native_balances(num_accounts: u32, user: &mut User) -> TestBalances {
+    pub(crate) fn get_native_balances(num_accounts: usize, user: &mut User) -> TestBalances {
         Self::get_asset_balances(AssetBase::native(), num_accounts, user)
     }
 
     pub(crate) fn get_asset_balances(
         asset: AssetBase,
-        num_accounts: u32,
+        num_accounts: usize,
         wallet: &mut User,
     ) -> TestBalances {
         let balances = (0..num_accounts)
@@ -30,13 +30,13 @@ impl TestBalances {
 }
 
 pub(crate) struct TransferInfo {
-    acc_idx_from: u32,
-    acc_idx_to: u32,
+    acc_idx_from: usize,
+    acc_idx_to: usize,
     amount: u64,
 }
 
 impl TransferInfo {
-    pub(crate) fn new(acc_idx_from: u32, acc_idx_to: u32, amount: u64) -> Self {
+    pub(crate) fn new(acc_idx_from: usize, acc_idx_to: usize, amount: u64) -> Self {
         TransferInfo {
             acc_idx_from,
             acc_idx_to,
@@ -51,13 +51,13 @@ impl TransferInfo {
 }
 
 pub(crate) struct BurnInfo {
-    burner_acc_idx: u32,
+    burner_acc_idx: usize,
     asset: AssetBase,
     amount: u64,
 }
 
 impl BurnInfo {
-    pub(crate) fn new(burner_acc_idx: u32, asset: AssetBase, amount: u64) -> Self {
+    pub(crate) fn new(burner_acc_idx: usize, asset: AssetBase, amount: u64) -> Self {
         BurnInfo {
             burner_acc_idx,
             asset,
@@ -73,11 +73,11 @@ impl BurnInfo {
 
 pub(crate) fn expected_balances_after_mine(
     balances: &TestBalances,
-    miner_idx: u32,
+    miner_idx: usize,
 ) -> TestBalances {
     let coinbase_value = 625_000_000;
     let mut new_balances = balances.clone();
-    new_balances.0[miner_idx as usize] += coinbase_value;
+    new_balances.0[miner_idx] += coinbase_value;
     new_balances
 }
 pub(crate) fn expected_balances_after_transfer(
@@ -87,8 +87,8 @@ pub(crate) fn expected_balances_after_transfer(
     let new_balances = transfers
         .iter()
         .fold(balances.clone(), |mut acc, transfer_info| {
-            acc.0[transfer_info.acc_idx_from as usize] -= transfer_info.amount;
-            acc.0[transfer_info.acc_idx_to as usize] += transfer_info.amount;
+            acc.0[transfer_info.acc_idx_from] -= transfer_info.amount;
+            acc.0[transfer_info.acc_idx_to] += transfer_info.amount;
             acc
         });
     new_balances
@@ -99,7 +99,7 @@ pub(crate) fn expected_balances_after_burn(
     burns: &[BurnInfo],
 ) -> TestBalances {
     let new_balances = burns.iter().fold(balances.clone(), |mut acc, burn_info| {
-        acc.0[burn_info.burner_acc_idx as usize] -= burn_info.amount;
+        acc.0[burn_info.burner_acc_idx] -= burn_info.amount;
         acc
     });
     new_balances
@@ -109,7 +109,7 @@ pub(crate) fn check_balances(
     asset: AssetBase,
     expected_balances: &TestBalances,
     user: &mut User,
-    num_accounts: u32,
+    num_accounts: usize,
 ) {
     let actual_balances = TestBalances::get_asset_balances(asset, num_accounts, user);
     assert_eq!(&actual_balances, expected_balances);
